[SHANIA][ADD/UPDATE] Modified Routes and Added Bottom Navigation Tab
