+++ app/(tabs)/leaderboard.tsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native'
+import React from 'react'
+
+const Leaderboard = () => {
+  return (
+    <View>
+      <Text>leaderboard</Text>
+    </View>
+  )
+}
+
+export default Leaderboard
