+++ app/(tabs)/profile.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Profile = () => {
+  return (
+    <View>
+      <Text>Profile</Text>
+    </View>
+  )
+}
+
+export default Profile
+
+const styles = StyleSheet.create({})
