+++ app/(tabs)/sandbox.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Sandbox = () => {
+  return (
+    <View>
+      <Text>Sandbox</Text>
+    </View>
+  )
+}
+
+export default Sandbox
+
+const styles = StyleSheet.create({})
