+++ app/(tabs)/setting.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Setting = () => {
+  return (
+    <View>
+      <Text>Setting</Text>
+    </View>
+  )
+}
+
+export default Setting
+
+const styles = StyleSheet.create({})
