+++ constants/icons.ts
@@ -1,0 +1,25 @@
+const homeOutline = require('../app/assets/icons/home-outline.png')
+const home = require('../app/assets/icons/home.png')
+const leaderboardOutline = require('../app/assets/icons/leaderboard-outline.png')
+const leaderboard = require('../app/assets/icons/leaderboard.png')
+const profileOutline = require('../app/assets/icons/profile-outline.png')
+const profile = require('../app/assets/icons/profile.png')
+const sandboxOutline = require('../app/assets/icons/sandbox-outline.png')
+const sandbox = require('../app/assets/icons/sandbox.png')
+const settingOutline = require('../app/assets/icons/setting-outline.png')
+const setting = require('../app/assets/icons/setting.png')
+
+const icons = {
+    homeOutline,
+    home,
+    leaderboardOutline,
+    leaderboard,
+    profileOutline,
+    profile,
+    sandboxOutline,
+    sandbox,
+    settingOutline,
+    setting
+}
+
+export default icons
